Refactoring
- 각 컴포넌트 별로 파일 분할
