+++ src/square.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export function Square(props) {
+  return (
+    <button
+      className="square"
+      onClick={props.onClick}
+    >
+      {props.value}
+    </button>
+  );
+}
